requirements-dev.txt
numpy<2.0.0
lightgbm>=4.0.0
pyomo==6.7.0
gurobipy
pytest-cov
IPython
nbsphinx
sphinx-rtd-theme
pre-commit